# executor: skip chunk evaluation in splitIntoGroups when unnecessary