<!-- Thank you for contributing to TiDB!
PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

When `firstRowDatums` are equal to `lastRowDatums` then there is no need to evaluate chunk. 

### What is changed and how it works?

Extract getting `firstRowDatums` and `lastRowDatums` from `evalGroupItemsAndResolveGroups` and call  `evalGroupItemsAndResolveGroups` only when `firstGroupKey` is different than `lastGroupKey`.

Benchmark for [BenchmarkAggGroupByNDV](https://github.com/pingcap/tidb/blob/a9dc72cf25589bc00260f1af4e5a17dac0f66474/executor/benchmark_test.go#L383):
```
name                                                                                                           old time/op  new time/op  delta
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10,_hasDistinct:false,_rows:10000000,_concurrency:4)-4         163ms ± 3%   104ms ± 5%  -35.86%  (p=0.008 n=5+5)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100,_hasDistinct:false,_rows:10000000,_concurrency:4)-4        161ms ± 1%   104ms ± 1%  -35.50%  (p=0.008 n=5+5)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4       162ms ± 1%   108ms ± 0%  -32.95%  (p=0.008 n=5+5)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4      162ms ± 0%   158ms ± 2%   -2.50%  (p=0.016 n=5+5)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:100000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4     168ms ± 1%   163ms ± 1%   -2.76%  (p=0.008 n=5+5)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:1000000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4    225ms ± 1%   227ms ± 3%     ~     (p=0.690 n=5+5)
AggGroupByNDV/(execType:stream,_aggFunc:sum,_ndv:10000000,_hasDistinct:false,_rows:10000000,_concurrency:4)-4   494ms ± 2%   487ms ± 3%     ~     (p=0.222 n=5+5)

```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test


### Release note <!-- bugfixes or new feature need a release note -->

- skip chunk evaluation in splitIntoGroups when unnecessary